// src/api/generated/models/listUsersRequest.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ListUsersRequestSortOrder } from './listUsersRequestSortOrder';

export interface ListUsersRequest {
  /** @minimum 1 */
  pageNum?: number;
  /** @minimum 1 */
  pageSize?: number;
  sortField?: string;
  sortOrder?: ListUsersRequestSortOrder;
  account?: string;
  createTimeStart?: string;
  createTimeEnd?: string;
}