// src/api/generated/models/pageResultUserVO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { UserVO } from './userVO';

export interface PageResultUserVO {
  list?: UserVO[];
  total?: string;
  pageNum?: number;
  pageSize?: number;
  totalPages?: number;
  hasNext?: boolean;
  hasPrevious?: boolean;
}