// src/api/generated/models/userVO.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface UserVO {
  id?: string;
  account?: string;
  role?: number;
  roleText?: string;
  nickname?: string;
  avatar?: string;
  profile?: string;
  createTime?: string;
}